+++ ruby/red-arrow/test/raw-records/test-struct-array.rb
@@ -41,7 +41,7 @@ module RawRecordsStructArrayTests
       [nil],
     ]
     target = build(:null, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_boolean
@@ -51,7 +51,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:boolean, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_int8
@@ -61,7 +61,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:int8, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_uint8
@@ -71,7 +71,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:uint8, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_int16
@@ -81,7 +81,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:int16, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_uint16
@@ -91,7 +91,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:uint16, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_int32
@@ -101,7 +101,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:int32, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_uint32
@@ -111,7 +111,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:uint32, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_int64
@@ -121,7 +121,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:int64, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_uint64
@@ -131,7 +131,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:uint64, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_float
@@ -141,7 +141,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:float, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_double
@@ -151,7 +151,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:double, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_binary
@@ -161,7 +161,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:binary, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_string
@@ -171,7 +171,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:string, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_date32
@@ -181,7 +181,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:date32, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_date64
@@ -191,7 +191,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:date64, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_timestamp_second
@@ -205,7 +205,7 @@ module RawRecordsStructArrayTests
                      unit: :second,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_timestamp_milli
@@ -219,7 +219,7 @@ module RawRecordsStructArrayTests
                      unit: :milli,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_timestamp_micro
@@ -233,7 +233,7 @@ module RawRecordsStructArrayTests
                      unit: :micro,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_timestamp_nano
@@ -247,7 +247,7 @@ module RawRecordsStructArrayTests
                      unit: :nano,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_time32_second
@@ -263,7 +263,7 @@ module RawRecordsStructArrayTests
                      unit: :second,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_time32_milli
@@ -279,7 +279,7 @@ module RawRecordsStructArrayTests
                      unit: :milli,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_time64_micro
@@ -295,7 +295,7 @@ module RawRecordsStructArrayTests
                      unit: :micro,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_time64_nano
@@ -311,7 +311,7 @@ module RawRecordsStructArrayTests
                      unit: :nano,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_decimal128
@@ -326,7 +326,7 @@ module RawRecordsStructArrayTests
                      scale: 2,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_decimal256
@@ -341,7 +341,7 @@ module RawRecordsStructArrayTests
                      scale: 2,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_month_interval
@@ -351,7 +351,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:month_interval, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_day_time_interval
@@ -361,7 +361,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:day_time_interval, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_month_day_nano_interval
@@ -371,7 +371,7 @@ module RawRecordsStructArrayTests
       [{"field" => nil}],
     ]
     target = build(:month_day_nano_interval, records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_list
@@ -388,7 +388,7 @@ module RawRecordsStructArrayTests
                      },
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_struct
@@ -408,7 +408,7 @@ module RawRecordsStructArrayTests
                      ],
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def test_map
@@ -423,7 +423,7 @@ module RawRecordsStructArrayTests
                      item: :boolean,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 
   def remove_union_field_names(records)
@@ -464,7 +464,7 @@ module RawRecordsStructArrayTests
                    },
                    records)
     assert_equal(remove_union_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_dense_union
@@ -491,7 +491,7 @@ module RawRecordsStructArrayTests
                    },
                    records)
     assert_equal(remove_union_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_dictionary
@@ -508,15 +508,43 @@ module RawRecordsStructArrayTests
                      ordered: false,
                    },
                    records)
-    assert_equal(records, target.raw_records)
+    assert_equal(records, actual_records(target))
   end
 end
+
+class EachRawRecordRecordBatchStructArrayTest < Test::Unit::TestCase
+  include RawRecordsStructArrayTests
+
+  def build(type, records)
+    Arrow::RecordBatch.new(build_schema(type), records)
+  end
+
+  def actual_records(target)
+    target.each_raw_record.to_a
+  end
+end
+
+class EachRawRecordTableStructArrayTest < Test::Unit::TestCase
+  include RawRecordsStructArrayTests
+
+  def build(type, records)
+    Arrow::Table.new(build_schema(type), records)
+  end
+  def actual_records(target)
+    target.each_raw_record.to_a
+  end
+end
+
 
 class RawRecordsRecordBatchStructArrayTest < Test::Unit::TestCase
   include RawRecordsStructArrayTests
 
   def build(type, records)
     Arrow::RecordBatch.new(build_schema(type), records)
+  end
+
+  def actual_records(target)
+    target.raw_records
   end
 end
 
@@ -525,5 +553,9 @@ class RawRecordsTableStructArrayTest < Test::Unit::TestCase
 
   def build(type, records)
     Arrow::Table.new(build_schema(type), records)
+  end
+
+  def actual_records(target)
+    target.raw_records
   end
 end
